--- .changeset/add-external-agent-authorizer.md
@@ -1,5 +1,0 @@
----
-'@truefoundry/trueforge': patch
----
-
-Add mode-specific external agent authorization providers backed by TrueFoundry agent permissions.
