+++ packages/trueforge/src/app.ts
@@ -24,7 +24,7 @@ import { createSettingsRouter } from './apis/settings';
 import { createAvailableSkillsRouter } from './apis/skills';
 import { createTurnsRouter } from './apis/turns';
 import type { Authenticator } from './auth/authenticator';
-import type { ExternalAuthorizer } from './auth/externalAuthorizer';
+import type { Authorizer } from './auth/authorizer';
 import { resolveRequestContext } from './auth/identity';
 import { createAdminAuthMiddleware, createAuthMiddleware } from './auth/middleware';
 import type { McpCatalog } from './catalog/McpCatalog';
@@ -200,8 +200,8 @@ export interface ServerDeps<TTransaction> {
   oidcClient: Configuration | undefined;
   /** Startup-selected authenticator; middleware is built from this once per app. */
   authenticator: Authenticator;
-  /** Startup-selected external agent authorization policy. */
-  externalAuthorizer: ExternalAuthorizer;
+  /** Startup-selected agent authorization policy. */
+  authorizer: Authorizer;
 }
 
 export function createServerApp<TTransaction>(deps: ServerDeps<TTransaction>) {
